pyproject.toml
[project]
name = "agno-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno>=2.2.10",
    "fastapi[standard]>=0.121.0",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "marimo>=0.18.1",
    "mcp>=1.21.0",
    "openai>=2.7.1",
    "openinference-instrumentation-agno>=0.1.23",
    "opentelemetry-api>=1.39.0",
    "opentelemetry-sdk>=1.39.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "tavily-python>=0.7.13",
    "uvicorn>=0.38.0",
]
